solve

+++ Difference.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main()
 {
-    int A,B,C,D;
+   long long  int A,B,C,D;
     cin>>A;
      cin>>B;
       cin>>C;
